move the component folder out of app folder
