+++ my-app/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono, Inter, Playfair_Display  } from "next/font/google";
 import "./globals.css";
-import Header from "./components/header"; 
-import { Footer } from './components/footer';
+import Header from "../components/header"; 
+import { Footer } from '../components/footer';
 
 const inter = Inter({
   subsets: ["latin"],
