+++ my-app/components/footer.tsx
